requirements.txt
fastapi
uvicorn
requests
scikit-learn
pandas
python-dotenv


